requirements.txt
numpy
scipy
scikit-learn
pprint
pyrebase
python-dotenv
tqdm